Implement Parcelable creation.
The implementation aims at changing as less code as possible from the original AutoValue so that it should be easier to maintain in the future.

The easiest way is to rely on writeValue and readValue with appropriate type cast instead of mapping every possible value. The loss in performance is negligible and the gain in code readability is humongous.